fix: ipad scrollbar v4
