+++ src/components/MainPage/ModalParams.tsx
@@ -1,28 +1,12 @@
 'use client';
 
 //import { useBodyScrollLock } from '@/hooks/useBodyScrollLock';
+import { useBodyScrollLockSimple } from '@/hooks/useBodyScrollLockSimple';
 import Link from 'next/link';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
 import { CloseIcon } from '../common/icons';
 import { Gratitude } from './Gratitude';
 import { PDFView } from './PdfView';
-
-const useBodyScrollLock = (shouldLock: boolean) => {
-  useEffect(() => {
-    const originalOverflow = window.getComputedStyle(document.body).overflow;
-    if (shouldLock) {
-      document.body.style.overflow = 'hidden';
-    } else {
-      document.body.style.overflow = originalOverflow;
-    }
-    return () => {
-      document.body.style.overflow = originalOverflow;
-    };
-  }, [shouldLock]);
-};
-
-export default useBodyScrollLock;
 
 export const ModalParams = () => {
   const params = useSearchParams();
@@ -32,7 +16,7 @@ export const ModalParams = () => {
   const document = params.get('document');
 
   //const bodyScrollLockRef = useBodyScrollLock(isPaymentSuccess || !!document);
-  useBodyScrollLock(isPaymentSuccess || !!document);
+  useBodyScrollLockSimple(isPaymentSuccess || !!document);
 
   const handlerShowModal = (
     e: React.MouseEvent<HTMLElement | SVGSVGElement>
@@ -46,11 +30,11 @@ export const ModalParams = () => {
   return isPaymentSuccess || document ? (
     <section
       onClick={handlerShowModal}
-      className="duration-250  fixed left-0 top-0 z-20 flex h-screen w-screen items-center justify-center bg-neutral-75  bg-opacity-30 backdrop-blur-2xl " //backdrop-filter
+      className="duration-250  fixed left-0 top-0 z-20 flex h-screen w-screen items-center justify-center bg-neutral-75  bg-opacity-30 backdrop-blur-2xl backdrop-filter"
       //ref={bodyScrollLockRef}
     >
       <div
-        className=" relative h-screen w-screen overflow-auto rounded-xl bg-white p-12 md:h-[80%] md:w-[80%] md:px-[6.85rem] md:py-[12.8rem]" //scrollbar
+        className="scrollbar relative h-screen w-screen overflow-auto rounded-xl bg-white p-12 md:h-[80%] md:w-[80%] md:px-[6.85rem] md:py-[12.8rem]"
         onClick={(e) => e.stopPropagation()}
       >
         <Link
